Handle gaps in line charts
