+++ .github/workflows/test_typescript.yml
@@ -31,6 +31,5 @@ jobs:
       - name: Type-checking
         run: yarn type-check
 
-      # Disabled because there are no TS tests currently.
-      #- name: Run tests
-      #  run: yarn workspaces foreach run test
+      - name: Run tests
+        run: yarn workspaces foreach run test
