+++ fiberplane-charts/src/Mondrian/generator/constants.ts
@@ -1,0 +1,2 @@
+export const BAR_PADDING = 0.2;
+export const BAR_PLUS_PADDING = 1 + BAR_PADDING;
