+++ fiberplane-charts/src/Mondrian/generator/generateLineChartFromTimeseries.ts
@@ -7,17 +7,19 @@ import type {
   TimeseriesSourceData,
 } from "../types";
 import {
+  calculateSmallestTimeInterval,
   calculateYAxisRange,
   createMetricBuckets,
   extendMinMax,
   getInitialMinMax,
   getTimeFromTimestamp,
   getXAxisFromTimeRange,
-  MinMax,
   normalizeAlongLinearAxis,
+  splitIntoContinuousLines,
 } from "./utils";
 import { identity } from "../../utils";
 import type { Metric, Timeseries } from "../../providerTypes";
+import type { MinMax } from "./types";
 
 export function generateLineChartFromTimeseries(
   input: TimeseriesSourceData,
@@ -31,10 +33,12 @@ export function generateLineChartFromTimeseries(
   const xAxis = getXAxisFromTimeRange(input.timeRange);
   const yAxis = calculateYAxisRange(buckets, identity);
 
+  const interval = calculateSmallestTimeInterval(buckets);
+
   const shapeLists: Array<ShapeList<Timeseries, Metric>> =
     input.timeseriesData.map((timeseries) => ({
       shapes: timeseries.visible
-        ? getShapes(timeseries.metrics, xAxis, yAxis)
+        ? getShapes(timeseries.metrics, xAxis, yAxis, interval)
         : [],
       source: timeseries,
     }));
@@ -46,6 +50,7 @@ function getShapes(
   metrics: Array<Metric>,
   xAxis: Axis,
   yAxis: Axis,
+  interval: number | null,
 ): Array<Shape<Metric>> {
   switch (metrics.length) {
     case 0:
@@ -56,16 +61,13 @@ function getShapes(
         ? []
         : [{ type: "point", ...getPointForMetric(metric, xAxis, yAxis) }];
     }
-    default:
-      // TODO: Implement gap detection: https://github.com/autometrics-dev/explorer/issues/35
-      return [
-        {
-          type: "line",
-          points: metrics.map((metric) =>
-            getPointForMetric(metric, xAxis, yAxis),
-          ),
-        },
-      ];
+    default: {
+      const lines = splitIntoContinuousLines(metrics, interval ?? undefined);
+      return lines.map((line) => ({
+        type: "line",
+        points: line.map((metric) => getPointForMetric(metric, xAxis, yAxis)),
+      }));
+    }
   }
 }
