+++ fiberplane-charts/src/Mondrian/generator/generateStackedLineChartFromTimeseries.ts
@@ -7,11 +7,13 @@ import type {
 } from "../types";
 import {
   calculateBucketsAndAxesForStackedChart,
+  calculateSmallestTimeInterval,
   getTimeFromTimestamp,
   normalizeAlongLinearAxis,
+  splitIntoContinuousLines,
 } from "./utils";
-import type { Metric, Timeseries } from "../../providerTypes";
 import { compact } from "../../utils";
+import type { Metric, Timeseries } from "../../providerTypes";
 
 type AxesAndBuckets = ReturnType<typeof calculateBucketsAndAxesForStackedChart>;
 
@@ -20,10 +22,12 @@ export function generateStackedLineChartFromTimeseries(
 ): AbstractChart<Timeseries, Metric> {
   const axesAndBuckets = calculateBucketsAndAxesForStackedChart(input);
 
+  const interval = calculateSmallestTimeInterval(axesAndBuckets.buckets);
+
   const shapeLists: Array<ShapeList<Timeseries, Metric>> =
     input.timeseriesData.map((timeseries) => ({
       shapes: timeseries.visible
-        ? getShapes(timeseries.metrics, axesAndBuckets)
+        ? getShapes(timeseries.metrics, axesAndBuckets, interval)
         : [],
       source: timeseries,
     }));
@@ -38,17 +42,15 @@ export function generateStackedLineChartFromTimeseries(
 function getShapes(
   metrics: Array<Metric>,
   axesAndBuckets: AxesAndBuckets,
+  interval: number | null,
 ): Array<Shape<Metric>> {
-  if (metrics.length === 0) {
-    return [];
-  }
-
-  // TODO: Implement gap detection: https://github.com/autometrics-dev/explorer/issues/35
-  const points = compact(
-    metrics.map((metric) => getPointForMetric(metric, axesAndBuckets)),
-  );
-
-  return [{ type: "area", points }];
+  const lines = splitIntoContinuousLines(metrics, interval ?? undefined);
+  return lines.map((line) => ({
+    type: "area",
+    points: compact(
+      line.map((metric) => getPointForMetric(metric, axesAndBuckets)),
+    ),
+  }));
 }
 
 function getPointForMetric(
