+++ fiberplane-charts/src/Mondrian/generator/utils/getInitialMinMax.ts
@@ -1,0 +1,8 @@
+import type { MinMax } from "../types";
+
+/**
+ * Returns the initial min-max based on a single value.
+ */
+export function getInitialMinMax(value: number): MinMax {
+  return [value, value];
+}
